+++ src/main/java/com/havanki/doppio/ContentTypeResolver.java
@@ -28,25 +28,24 @@ import java.util.List;
  */
 public class ContentTypeResolver {
 
-  private static final String GMI_SUFFIX = ".gmi";
-  private static final String GEMINI_SUFFIX = ".gemini";
-  private static final List<String> GEMINI_SUFFIXES = List.of(
-    GMI_SUFFIX,
-    GEMINI_SUFFIX
-  );
-
   /**
    * The default content type to return when all else fails.
    */
   public static final String DEFAULT_CONTENT_TYPE = "application/octet-stream";
 
   private static final FileNameMap FILE_NAME_MAP =
-      URLConnection.getFileNameMap();
+    URLConnection.getFileNameMap();
+
+  private final List<String> textGeminiSuffixes;
+
+  public ContentTypeResolver(List<String> textGeminiSuffixes) {
+    this.textGeminiSuffixes = textGeminiSuffixes;
+  }
 
   /**
    * Gets the content type for a file, based on just its filename. This relies
-   * mostly on Java's built-in MIME type detection, but adds support for .gmi
-   * and .gemini as text/gemini.<p>
+   * mostly on Java's built-in MIME type detection, but adds support for
+   * text/gemini.<p>
    *
    * If none of the usual detection techniques work, then this method returns
    * {@link #DEFAULT_CONTENT_TYPE}.
@@ -55,20 +54,10 @@ public class ContentTypeResolver {
    * @return          content type
    */
   public String getContentTypeFor(String fileName) {
-    if (fileName.endsWith(GMI_SUFFIX) ||
-        fileName.endsWith(GEMINI_SUFFIX)) {
+    if (textGeminiSuffixes.stream().anyMatch(s -> fileName.endsWith(s))) {
       return "text/gemini";
     }
     String contentType = FILE_NAME_MAP.getContentTypeFor(fileName);
     return contentType != null ? contentType : DEFAULT_CONTENT_TYPE;
-  }
-
-  /**
-   * Gets the filename suffixes recognized as text/gemini.
-   *
-   * @return text/gemini filename suffixes
-   */
-  public Iterable<String> getGeminiSuffixes() {
-    return GEMINI_SUFFIXES;
   }
 }